requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Asterisk AGI
pyst2==0.5.1
asterisk-agi==0.1.5

# Speech recognition
deepgram-sdk==3.2.0
openai-whisper==20231117
soundfile==0.12.1
pydub==0.25.1

# Audio processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1

# AI and NLP
openai==1.3.7
anthropic==0.7.7
langchain==0.1.0
langchain-openai==0.0.2

# WebSocket for real-time communication
websockets==12.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
loguru==0.7.2

# Database (for call records and leads)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis for state management
redis==5.0.1
hiredis==2.2.3

# Utilities
python-multipart==0.0.6
pytz==2023.3
phonenumbers==8.13.26
